Move watcher job-runner
